requirements.txt
colorama~=0.4.6
web3~=6.8.0
PyYAML~=6.0.1
loguru~=0.7.2
pydantic~=2.5.2
inquirer~=3.1.3
art~=6.1